Fix a nasty bug which occured sometimes when there was transient loss of internet connectivity.
All bookmarks are supposed to be fetched only when the program is being run for the first time. When the program is being run for the first time, lastUpdate.txt file doesn't exist and there is an except block handling this condition. But that except block was also being executed when there was no internet connectivity. In the rare case when internet connectivity was reestablished just when the except block started executing, the program would incorrectly fetch all bookmarks.

I moved the code which was fetching the bookmarks to a different try-except block to fix this bug.